[form-factors] Move ``BCL2008`` parametrization to separate files
